requirements.txt
Flask==2.2.5
numpy==1.26.4
Pillow==11.3.0
gunicorn==23.0.0
